rename project and repackage
