--- airsync-client/src/main/kotlin/ffc/airsync/client/log/debug.kt
@@ -1,8 +1,0 @@
-package ffc.airsync.client.log
-
-val debug = System.getenv("FFC_DEBUG")
-
-inline fun printDebug(infoDebug: String) {
-    if (debug == null)
-        println(infoDebug)
-}
