--- airsync-client/src/main/resources/main.css
@@ -1,54 +1,0 @@
-/**
- * Copyright 2015 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-html, body {
-  font-family: 'Roboto', 'Helvetica', sans-serif;
-}
-a {
-  text-decoration: none;
-}
-li a {
-  text-decoration: underline;
-  color: #0288d1;
-}
-.mdl-grid {
-  max-width: 1024px;
-  margin: auto;
-}
-.mdl-layout__header-row {
-  padding: 0;
-}
-.quickstart-user-details-container {
-  margin-top: 20px;
-  line-height: 25px;
-}
-#quickstart-sign-in-status {
-  font-weight: bold;
-}
-pre {
-  overflow-x: scroll;
-  line-height: 18px;
-}
-code {
-  white-space: pre-wrap;
-  word-break: break-all;
-}
-h3 {
-  background: url('/firebase-logo.png') no-repeat;
-  background-size: 40px;
-  padding-left: 50px;
-  color: white;
-}
