+++ airsync/src/main/kotlin/ffc/airsync/utils/debug.kt
@@ -1,0 +1,8 @@
+package ffc.airsync.utils
+
+val debug = System.getenv("FFC_DEBUG")
+
+inline fun printDebug(infoDebug: String) {
+    if (debug == null)
+        println(infoDebug)
+}
