+++ implement/airsync-local-web/src/main/resources/firebase-messaging-sw.js
@@ -1,0 +1,23 @@
+importScripts('https://www.gstatic.com/firebasejs/4.13.0/firebase-app.js');
+importScripts('https://www.gstatic.com/firebasejs/4.13.0/firebase-messaging.js');
+
+firebase.initializeApp({
+    'messagingSenderId': '656545465'
+});
+
+
+const messaging = firebase.messaging();
+
+
+messaging.setBackgroundMessageHandler(function (payload) {
+    console.log('[firebase-messaging-sw.js] Received background message ', payload);
+    // Customize notification here
+    var notificationTitle = 'Background Message Title';
+    var notificationOptions = {
+        body: 'Background Message body.',
+        icon: '/firebase-logo.png'
+    };
+
+    return self.registration.showNotification(notificationTitle,
+        notificationOptions);
+});
