+++ implement/airsync-local-web/src/main/resources/manifest.json
